Renamed the consts file
